Split SvgMoveToSegment
